feat: testing productsDB

+++ 39/internal/entity/product.go
@@ -12,17 +12,17 @@ var (
 	ErrInvalidID        = errors.New("invalid ID")
 	ErrNameIsRequiered  = errors.New("name is required")
 	ErrPriceIsRequiered = errors.New("price is required")
-	ErrInvalidPrice		 	= errors.New("invalid price")
+	ErrInvalidPrice     = errors.New("invalid price")
 )
 
 type Product struct {
 	ID        entity.ID `json:"id"`
 	Name      string    `json:"name"`
-	Price     int       `json:"price"`
+	Price     float64   `json:"price"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
-func NewProduct(name string, price int) (*Product, error) {
+func NewProduct(name string, price float64) (*Product, error) {
 	product := &Product{
 		ID:        entity.NewID(),
 		Name:      name,
